Remove daemon.Visor
