+++ src/daemon/daemon.go
@@ -8,8 +8,10 @@ import (
 	"time"
 
 	"github.com/skycoin/skycoin/src/cipher"
+	"github.com/skycoin/skycoin/src/coin"
 	"github.com/skycoin/skycoin/src/daemon/gnet"
 	"github.com/skycoin/skycoin/src/daemon/pex"
+	"github.com/skycoin/skycoin/src/visor"
 	"github.com/skycoin/skycoin/src/visor/dbutil"
 
 	"github.com/skycoin/skycoin/src/util/elapse"
@@ -64,7 +66,7 @@ type Config struct {
 	Pool     PoolConfig
 	Pex      pex.Config
 	Gateway  GatewayConfig
-	Visor    VisorConfig
+	Visor    visor.Config
 }
 
 // NewConfig returns a Config with defaults set
@@ -75,7 +77,7 @@ func NewConfig() Config {
 		Pex:      pex.NewConfig(),
 		Gateway:  NewGatewayConfig(),
 		Messages: NewMessagesConfig(),
-		Visor:    NewVisorConfig(),
+		Visor:    visor.NewVisorConfig(),
 	}
 }
 
@@ -104,15 +106,12 @@ func (cfg *Config) preprocess() Config {
 		config.Pex.Disabled = true
 		config.Daemon.DisableIncomingConnections = true
 		config.Daemon.DisableOutgoingConnections = true
-		config.Visor.DisableNetworking = true
 	} else {
 		if config.Daemon.DisableIncomingConnections {
 			logger.Info("Incoming connections are disabled.")
 		}
 		if config.Daemon.DisableOutgoingConnections {
 			logger.Info("Outgoing connections are disabled.")
-			// Visor only makes outgoing connections
-			config.Visor.DisableNetworking = true
 		}
 	}
 
@@ -121,7 +120,7 @@ func (cfg *Config) preprocess() Config {
 
 // DaemonConfig configuration for the Daemon
 type DaemonConfig struct { // nolint: golint
-	// Application version. TODO -- manage version better
+	// Protocol version. TODO -- manage version better
 	Version int32
 	// IP Address to serve on. Leave empty for automatic assignment
 	Address string
@@ -156,27 +155,48 @@ type DaemonConfig struct { // nolint: golint
 	LocalhostOnly bool
 	// Log ping and pong messages
 	LogPings bool
+	// How often to request blocks from peers
+	BlocksRequestRate time.Duration
+	// How often to announce our blocks to peers
+	BlocksAnnounceRate time.Duration
+	// How many blocks to respond with to a GetBlocksMessage
+	BlocksResponseCount uint64
+	// Max announce txns hash number
+	MaxTxnAnnounceNum int
+	// How often new blocks are created by the signing node, in seconds
+	BlockCreationInterval uint64
+	// How often to check the unconfirmed pool for transactions that become valid
+	UnconfirmedRefreshRate time.Duration
+	// How often to remove transactions that become permanently invalid from the unconfirmed pool
+	UnconfirmedRemoveInvalidRate time.Duration
 }
 
 // NewDaemonConfig creates daemon config
 func NewDaemonConfig() DaemonConfig {
 	return DaemonConfig{
-		Version:                    2,
-		Address:                    "",
-		Port:                       6677,
-		OutgoingRate:               time.Second * 5,
-		PrivateRate:                time.Second * 5,
-		OutgoingMax:                8,
-		PendingMax:                 8,
-		IntroductionWait:           time.Second * 30,
-		CullInvalidRate:            time.Second * 3,
-		FlushAnnouncedTxnsRate:     time.Second * 3,
-		IPCountsMax:                3,
-		DisableNetworking:          false,
-		DisableOutgoingConnections: false,
-		DisableIncomingConnections: false,
-		LocalhostOnly:              false,
-		LogPings:                   true,
+		Version:                      2,
+		Address:                      "",
+		Port:                         6677,
+		OutgoingRate:                 time.Second * 5,
+		PrivateRate:                  time.Second * 5,
+		OutgoingMax:                  8,
+		PendingMax:                   8,
+		IntroductionWait:             time.Second * 30,
+		CullInvalidRate:              time.Second * 3,
+		FlushAnnouncedTxnsRate:       time.Second * 3,
+		IPCountsMax:                  3,
+		DisableNetworking:            false,
+		DisableOutgoingConnections:   false,
+		DisableIncomingConnections:   false,
+		LocalhostOnly:                false,
+		LogPings:                     true,
+		BlocksRequestRate:            time.Second * 60,
+		BlocksAnnounceRate:           time.Second * 60,
+		BlocksResponseCount:          20,
+		MaxTxnAnnounceNum:            16,
+		BlockCreationInterval:        10,
+		UnconfirmedRefreshRate:       time.Minute,
+		UnconfirmedRemoveInvalidRate: time.Minute,
 	}
 }
 
@@ -190,9 +210,14 @@ type Daemon struct {
 	Pool     *Pool
 	Pex      *pex.Pex
 	Gateway  *Gateway
-	Visor    *Visor
+	Visor    *visor.Visor
 
 	DefaultConnections []string
+
+	// Cache of announced transactions that are flushed to the database periodically
+	announcedTxns *announcedTxnsCache
+	// Cache of reported peer blockchain heights
+	heights *peerBlockchainHeights
 
 	// Separate index of outgoing connections. The pool aggregates all
 	// connections.
@@ -221,8 +246,6 @@ type Daemon struct {
 	ipCounts *IPCount
 	// Message handling queue
 	messageEvents chan MessageEvent
-	// Cache of announced transactions that are flushed to the database periodically
-	announcedTxns *announcedTxnsCache
 	// quit channel
 	quit chan struct{}
 	// done channel
@@ -234,7 +257,7 @@ type Daemon struct {
 // NewDaemon returns a Daemon with primitives allocated
 func NewDaemon(config Config, db *dbutil.DB, defaultConns []string) (*Daemon, error) {
 	config = config.preprocess()
-	vs, err := NewVisor(config.Visor, db)
+	vs, err := visor.NewVisor(config.Visor, db)
 	if err != nil {
 		return nil, err
 	}
@@ -250,7 +273,10 @@ func NewDaemon(config Config, db *dbutil.DB, defaultConns []string) (*Daemon, er
 		Pex:      pex,
 		Visor:    vs,
 
-		DefaultConnections: defaultConns, //passed in from top level
+		DefaultConnections: defaultConns,
+
+		announcedTxns: newAnnouncedTxnsCache(),
+		heights:       newPeerBlockchainHeights(),
 
 		expectingIntroductions: NewExpectIntroductions(),
 		connectionMirrors:      NewConnectionMirrors(),
@@ -265,7 +291,6 @@ func NewDaemon(config Config, db *dbutil.DB, defaultConns []string) (*Daemon, er
 		outgoingConnections: NewOutgoingConnections(config.Daemon.OutgoingMax),
 		pendingConnections:  NewPendingConnections(config.Daemon.PendingMax),
 		messageEvents:       make(chan MessageEvent, config.Pool.EventChannelSize),
-		announcedTxns:       newAnnouncedTxnsCache(),
 		quit:                make(chan struct{}),
 		done:                make(chan struct{}),
 	}
@@ -330,7 +355,7 @@ func (dm *Daemon) Run() error {
 	defer logger.Info("Daemon closed")
 	defer close(dm.done)
 
-	if err := dm.Visor.v.Init(); err != nil {
+	if err := dm.Visor.Init(); err != nil {
 		logger.WithError(err).Error("visor.Visor.Init failed")
 		return err
 	}
@@ -363,23 +388,20 @@ func (dm *Daemon) Run() error {
 		}
 	}()
 
-	// TODO -- run blockchain stuff in its own goroutine
-	blockInterval := time.Duration(dm.Visor.Config.Config.BlockCreationInterval)
-	// blockchainBackupTicker := time.Tick(self.Visor.Config.BlockchainBackupRate)
+	blockInterval := time.Duration(dm.Config.BlockCreationInterval)
 	blockCreationTicker := time.NewTicker(time.Second * blockInterval)
-	if !dm.Visor.Config.Config.IsMaster {
+	if !dm.Visor.Config.IsMaster {
 		blockCreationTicker.Stop()
 	}
 
-	unconfirmedRefreshTicker := time.Tick(dm.Visor.Config.Config.UnconfirmedRefreshRate)
-	unconfirmedRemoveInvalidTicker := time.Tick(dm.Visor.Config.Config.UnconfirmedRemoveInvalidRate)
-	blocksRequestTicker := time.Tick(dm.Visor.Config.BlocksRequestRate)
-	blocksAnnounceTicker := time.Tick(dm.Visor.Config.BlocksAnnounceRate)
+	unconfirmedRefreshTicker := time.Tick(dm.Config.UnconfirmedRefreshRate)
+	unconfirmedRemoveInvalidTicker := time.Tick(dm.Config.UnconfirmedRemoveInvalidRate)
+	blocksRequestTicker := time.Tick(dm.Config.BlocksRequestRate)
+	blocksAnnounceTicker := time.Tick(dm.Config.BlocksAnnounceRate)
 
 	privateConnectionsTicker := time.Tick(dm.Config.PrivateRate)
 	cullInvalidTicker := time.Tick(dm.Config.CullInvalidRate)
 	outgoingConnectionsTicker := time.Tick(dm.Config.OutgoingRate)
-	// clearOldPeersTicker := time.Tick(dm.Peers.Config.CullRate)
 	requestPeersTicker := time.Tick(dm.Pex.Config.RequestRate)
 	clearStaleConnectionsTicker := time.Tick(dm.Pool.Config.ClearStaleRate)
 	idleCheckTicker := time.Tick(dm.Pool.Config.IdleCheckRate)
@@ -518,7 +540,11 @@ loop:
 		case <-flushAnnouncedTxnsTicker:
 			elapser.Register("flushAnnouncedTxnsTicker")
 			txns := dm.announcedTxns.flush()
-			dm.Visor.SetTxnsAnnounced(txns)
+
+			if err := dm.Visor.SetTxnsAnnounced(txns); err != nil {
+				logger.WithError(err).Error("Failed to set unconfirmed txn announce time")
+				return err
+			}
 
 		case m := <-dm.messageEvents:
 			// Message handlers
@@ -537,8 +563,8 @@ loop:
 		case <-blockCreationTicker.C:
 			// Create blocks, if master chain
 			elapser.Register("blockCreationTicker.C")
-			if dm.Visor.Config.Config.IsMaster {
-				sb, err := dm.Visor.CreateAndPublishBlock(dm.Pool)
+			if dm.Visor.Config.IsMaster {
+				sb, err := dm.CreateAndPublishBlock()
 				if err != nil {
 					logger.Errorf("Failed to create and publish block: %v", err)
 					continue
@@ -558,7 +584,7 @@ loop:
 				continue
 			}
 			// Announce these transactions
-			dm.Visor.AnnounceTxns(dm.Pool, validTxns)
+			dm.AnnounceTxns(validTxns)
 
 		case <-unconfirmedRemoveInvalidTicker:
 			elapser.Register("unconfirmedRemoveInvalidTicker")
@@ -574,11 +600,11 @@ loop:
 
 		case <-blocksRequestTicker:
 			elapser.Register("blocksRequestTicker")
-			dm.Visor.RequestBlocks(dm.Pool)
+			dm.RequestBlocks()
 
 		case <-blocksAnnounceTicker:
 			elapser.Register("blocksAnnounceTicker")
-			dm.Visor.AnnounceBlocks(dm.Pool)
+			dm.AnnounceBlocks()
 
 		case err = <-errC:
 			break loop
@@ -857,7 +883,7 @@ func (dm *Daemon) onDisconnect(e DisconnectEvent) {
 
 	dm.outgoingConnections.Remove(e.Addr)
 	dm.expectingIntroductions.Remove(e.Addr)
-	dm.Visor.RemoveConnection(e.Addr)
+	dm.heights.Remove(e.Addr)
 	dm.removeIPCount(e.Addr)
 	dm.removeConnectionMirror(e.Addr)
 }
@@ -969,38 +995,232 @@ func (dm *Daemon) handleMessageSendResult(r gnet.SendResult) {
 	}
 }
 
-type announcedTxnsCache struct {
-	sync.Mutex
-	cache map[cipher.SHA256]int64
-}
-
-func newAnnouncedTxnsCache() *announcedTxnsCache {
-	return &announcedTxnsCache{
-		cache: make(map[cipher.SHA256]int64),
-	}
-}
-
-func (c *announcedTxnsCache) add(txns []cipher.SHA256) {
-	c.Lock()
-	defer c.Unlock()
-
-	t := utc.Now().UnixNano()
-	for _, txn := range txns {
-		c.cache[txn] = t
-	}
-}
-
-func (c *announcedTxnsCache) flush() map[cipher.SHA256]int64 {
-	c.Lock()
-	defer c.Unlock()
-
-	if len(c.cache) == 0 {
+// RequestBlocks Sends a GetBlocksMessage to all connections
+func (dm *Daemon) RequestBlocks() error {
+	if dm.Config.DisableOutgoingConnections {
 		return nil
 	}
 
-	cache := c.cache
+	headSeq, ok, err := dm.Visor.HeadBkSeq()
+	if err != nil {
+		return err
+	}
+	if !ok {
+		return errors.New("Cannot request blocks, there is no head block")
+	}
 
-	c.cache = make(map[cipher.SHA256]int64)
+	m := NewGetBlocksMessage(headSeq, dm.Config.BlocksResponseCount)
 
-	return cache
+	err = dm.Pool.Pool.BroadcastMessage(m)
+	if err != nil {
+		logger.Debugf("Broadcast GetBlocksMessage failed: %v", err)
+	}
+
+	return err
+}
+
+// AnnounceBlocks sends an AnnounceBlocksMessage to all connections
+func (dm *Daemon) AnnounceBlocks() error {
+	if dm.Config.DisableOutgoingConnections {
+		return nil
+	}
+
+	headSeq, ok, err := dm.Visor.HeadBkSeq()
+	if err != nil {
+		return err
+	}
+	if !ok {
+		return errors.New("Cannot announce blocks, there is no head block")
+	}
+
+	m := NewAnnounceBlocksMessage(headSeq)
+
+	err = dm.Pool.Pool.BroadcastMessage(m)
+	if err != nil {
+		logger.Debugf("Broadcast AnnounceBlocksMessage failed: %v", err)
+	}
+
+	return err
+}
+
+// AnnounceAllTxns announces local unconfirmed transactions
+func (dm *Daemon) AnnounceAllTxns() error {
+	if dm.Config.DisableOutgoingConnections {
+		return nil
+	}
+
+	// Get local unconfirmed transaction hashes.
+	hashes, err := dm.Visor.GetAllValidUnconfirmedTxHashes()
+	if err != nil {
+		return err
+	}
+
+	// Divide hashes into multiple sets of max size
+	hashesSet := divideHashes(hashes, dm.Config.MaxTxnAnnounceNum)
+
+	for _, hs := range hashesSet {
+		m := NewAnnounceTxnsMessage(hs)
+		if err = dm.Pool.Pool.BroadcastMessage(m); err != nil {
+			break
+		}
+	}
+
+	if err != nil {
+		logger.Debugf("Broadcast AnnounceTxnsMessage failed, err: %v", err)
+	}
+
+	return err
+}
+
+func divideHashes(hashes []cipher.SHA256, n int) [][]cipher.SHA256 {
+	if len(hashes) == 0 {
+		return [][]cipher.SHA256{}
+	}
+
+	var j int
+	var hashesArray [][]cipher.SHA256
+
+	if len(hashes) > n {
+		for i := range hashes {
+			if len(hashes[j:i]) == n {
+				hs := make([]cipher.SHA256, n)
+				copy(hs, hashes[j:i])
+				hashesArray = append(hashesArray, hs)
+				j = i
+			}
+		}
+	}
+
+	hs := make([]cipher.SHA256, len(hashes)-j)
+	copy(hs, hashes[j:])
+	hashesArray = append(hashesArray, hs)
+	return hashesArray
+}
+
+// AnnounceTxns announces given transaction hashes.
+func (dm *Daemon) AnnounceTxns(txns []cipher.SHA256) error {
+	if dm.Config.DisableOutgoingConnections {
+		return nil
+	}
+
+	if len(txns) == 0 {
+		return nil
+	}
+
+	m := NewAnnounceTxnsMessage(txns)
+
+	err := dm.Pool.Pool.BroadcastMessage(m)
+	if err != nil {
+		logger.Debugf("Broadcast AnnounceTxnsMessage failed: %v", err)
+	}
+
+	return err
+}
+
+// RequestBlocksFromAddr sends a GetBlocksMessage to one connected address
+func (dm *Daemon) RequestBlocksFromAddr(addr string) error {
+	if dm.Config.DisableOutgoingConnections {
+		return errors.New("Outgoing connections disabled")
+	}
+
+	headSeq, ok, err := dm.Visor.HeadBkSeq()
+	if err != nil {
+		return err
+	}
+	if !ok {
+		return errors.New("Cannot request blocks from addr, there is no head block")
+	}
+
+	m := NewGetBlocksMessage(headSeq, dm.Config.BlocksResponseCount)
+
+	return dm.Pool.Pool.SendMessage(addr, m)
+}
+
+// InjectBroadcastTransaction injects transaction to the unconfirmed pool and broadcasts it.
+// If the transaction violates either hard or soft constraints, it is not broadcast.
+// This method is to be used by user-initiated transaction injections.
+// For transactions received over the network, use InjectTransaction and check the result to
+// decide on repropagation.
+func (dm *Daemon) InjectBroadcastTransaction(txn coin.Transaction) error {
+	if _, err := dm.Visor.InjectTransactionStrict(txn); err != nil {
+		return err
+	}
+
+	return dm.broadcastTransaction(txn)
+}
+
+// ResendUnconfirmedTxns resends all unconfirmed transactions and returns the hashes that were successfully rebroadcast
+func (dm *Daemon) ResendUnconfirmedTxns() ([]cipher.SHA256, error) {
+	if dm.Config.DisableOutgoingConnections {
+		return nil, nil
+	}
+
+	txns, err := dm.Visor.GetAllUnconfirmedTxns()
+	if err != nil {
+		return nil, err
+	}
+
+	var txids []cipher.SHA256
+	for i := range txns {
+		logger.Debugf("Rebroadcast tx %s", txns[i].Hash().Hex())
+		if err := dm.broadcastTransaction(txns[i].Txn); err == nil {
+			txids = append(txids, txns[i].Txn.Hash())
+		}
+	}
+
+	return txids, nil
+}
+
+// broadcastTransaction broadcasts a single transaction to all peers.
+func (dm *Daemon) broadcastTransaction(t coin.Transaction) error {
+	if dm.Config.DisableOutgoingConnections {
+		return nil
+	}
+
+	m := NewGiveTxnsMessage(coin.Transactions{t})
+	l, err := dm.Pool.Pool.Size()
+	if err != nil {
+		return err
+	}
+
+	logger.Debugf("Broadcasting GiveTxnsMessage to %d conns", l)
+
+	err = dm.Pool.Pool.BroadcastMessage(m)
+	if err != nil {
+		logger.Errorf("Broadcast GivenTxnsMessage failed: %v", err)
+	}
+
+	return err
+}
+
+// CreateAndPublishBlock creates a block from unconfirmed transactions and sends it to the network.
+// Will panic if not running as a master chain.
+// Will not create a block if outgoing connections are disabled.
+// If the block was created but the broadcast failed, the error will be non-nil but the
+// SignedBlock value will not be empty.
+// TODO -- refactor this method -- it should either always create a block and maybe broadcast it,
+// or use a database transaction to rollback block publishing if broadcast failed (however, this will cause a slow DB write)
+func (dm *Daemon) CreateAndPublishBlock() (*coin.SignedBlock, error) {
+	if dm.Config.DisableOutgoingConnections {
+		return nil, errors.New("Outgoing connections disabled")
+	}
+
+	sb, err := dm.Visor.CreateAndExecuteBlock()
+	if err != nil {
+		return nil, err
+	}
+
+	err = dm.broadcastBlock(sb)
+
+	return &sb, err
+}
+
+// Sends a signed block to all connections.
+func (dm *Daemon) broadcastBlock(sb coin.SignedBlock) error {
+	if dm.Config.DisableOutgoingConnections {
+		return nil
+	}
+
+	m := NewGiveBlocksMessage([]coin.SignedBlock{sb})
+	return dm.Pool.Pool.BroadcastMessage(m)
 }
