+++ src/daemon/daemon_test.go
@@ -1,1 +1,105 @@
 package daemon
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+
+	"github.com/skycoin/skycoin/src/cipher"
+)
+
+func TestDivideHashes(t *testing.T) {
+	hashes := make([]cipher.SHA256, 10)
+	for i := 0; i < 10; i++ {
+		hashes[i] = cipher.SumSHA256(cipher.RandByte(512))
+	}
+
+	testCases := []struct {
+		name  string
+		init  []cipher.SHA256
+		n     int
+		array [][]cipher.SHA256
+	}{
+		{
+			"has one odd",
+			hashes[:],
+			3,
+			[][]cipher.SHA256{
+				[]cipher.SHA256{
+					hashes[0],
+					hashes[1],
+					hashes[2],
+				},
+				[]cipher.SHA256{
+					hashes[3],
+					hashes[4],
+					hashes[5],
+				},
+				[]cipher.SHA256{
+					hashes[6],
+					hashes[7],
+					hashes[8],
+				},
+				[]cipher.SHA256{
+					hashes[9],
+				},
+			},
+		},
+		{
+			"only one value",
+			hashes[:1],
+			1,
+			[][]cipher.SHA256{
+				[]cipher.SHA256{
+					hashes[0],
+				},
+			},
+		},
+		{
+			"empty value",
+			hashes[:0],
+			0,
+			[][]cipher.SHA256{},
+		},
+		{
+			"with 3 value",
+			hashes[:3],
+			3,
+			[][]cipher.SHA256{
+				[]cipher.SHA256{
+					hashes[0],
+					hashes[1],
+					hashes[2],
+				},
+			},
+		},
+		{
+			"with 8 value",
+			hashes[:8],
+			3,
+			[][]cipher.SHA256{
+				[]cipher.SHA256{
+					hashes[0],
+					hashes[1],
+					hashes[2],
+				},
+				[]cipher.SHA256{
+					hashes[3],
+					hashes[4],
+					hashes[5],
+				},
+				[]cipher.SHA256{
+					hashes[6],
+					hashes[7],
+				},
+			},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			rlt := divideHashes(tc.init, tc.n)
+			require.Equal(t, tc.array, rlt)
+		})
+	}
+}
